polish: bug fixes + ruff modernization
BUG FIXES (real behavior regressions)
  - /metrics gauges now reflect runtime state, not startup config.
    Before: caveman_enabled stayed at startup value forever even after
    POST /settings. After: gauge tracks _runtime live; verified by
    flipping caveman 1->0->1 via curl.
  - PayloadCompressor.compress_request_payload now exposes
    force_enabled flag so callers that have already consulted the
    runtime gate (server.py proxy path) don't get short-circuited by
    the stale-settings static check.
  - PayloadCompressor.compress_response_payload accepts force_enabled
    on the same principle; server.py threads bool(rt['output_compression'])
    through.
  - preview.py now accepts input_compression/json_aware/lsh runtime
    overrides so the /preview endpoint reflects what would happen at
    request time, not what the static config says.
  - server.py:_should_transform_json_response: dropped the redundant
    static settings.input_compression_enabled check; the live runtime
    gate is already authoritative.

DEPRECATION FIXES (UP* modernizations - automated)
  - UP035: typing.Dict/List/Tuple/Optional -> dict/list/tuple/X | None
    across 12 modules
  - UP037: removed quoted annotations no longer needed in 3.11+
  - UP012: removed redundant .encode('utf-8') (default already utf-8)
  - UP045: stripped Optional[X] -> X | None wherever still present
  - RUF021: parens around mixed-precedence boolean expressions
  - RUF022: sorted __all__

CORRECTNESS HARDENING (B905 strict-zip)
  - All zip() calls now use strict=True or strict=False explicitly.
    6 src/ sites + 2 tests/ sites. Catches silent same-length
    assumptions that would otherwise truncate.

LINT CLEANUP
  - F401: removed unused imports in providers/__init__.py (Mapping)
    and lsh_dedupe.py (Iterable, Mapping)
  - RUF046: dropped int() casts where result is already int
    (audit.py timestamps, token_estimate.py division)
  - RUF059: replaced unused unpacked tuple variables with _ in
    5 test files

VERIFICATION
  - 708 pytest pass, 10 consecutive runs stable
  - ruff check src/ tests/: All checks passed
  - Live smoke test: /metrics gauge flips correctly on toggle,
    all 13 dashboard toggles bind, 4 providers + 10 routes,
    /healthz ok, dashboard renders 60KB, /v1/messages 401 without
    auth

+++ src/middleout_proxy/audit.py
@@ -31,7 +31,7 @@ def _bin_index(latency_ms: float) -> int:
 def _quantile_from_hist(counts: list[int], total: int, q: float) -> float:
     if total <= 0:
         return 0.0
-    target = max(1, int(math.ceil(q * total)))
+    target = max(1, math.ceil(q * total))
     cum = 0
     for i, c in enumerate(counts):
         cum += c
